TW-2265: reuse shared onStorageChanged helper
