+++ src/app/hooks/use-web-widgets-sync.ts
@@ -3,12 +3,12 @@ import { useEffect } from 'react';
 import { dispatch } from 'app/store';
 import { setWebWidgetsTokenInsightEnabledAction } from 'app/store/settings/actions';
 import { WEB_WIDGETS_TOKEN_INSIGHT_ENABLED } from 'lib/constants';
-import { onStorageKey } from 'lib/web-widgets/storage';
+import { onStorageChanged } from 'lib/storage';
 
 export const useWebWidgetsSync = () => {
   useEffect(
     () =>
-      onStorageKey<boolean>(WEB_WIDGETS_TOKEN_INSIGHT_ENABLED, value => {
+      onStorageChanged<boolean>(WEB_WIDGETS_TOKEN_INSIGHT_ENABLED, value => {
         if (typeof value === 'boolean') dispatch(setWebWidgetsTokenInsightEnabledAction(value));
       }),
     []
