+++ src/lib/web-widgets/pill/card/permit.ts
@@ -1,7 +1,7 @@
 import browser from 'webextension-polyfill';
 
 import { WEB_WIDGETS_LOCAL_AD_PERMIT } from 'lib/constants';
-import { onStorageKey } from 'lib/web-widgets/storage';
+import { onStorageChanged } from 'lib/storage';
 
 export const readAdPermit = async (): Promise<boolean> => {
   const stored = await browser.storage.local.get(WEB_WIDGETS_LOCAL_AD_PERMIT);
@@ -11,6 +11,6 @@ export const readAdPermit = async (): Promise<boolean> => {
 export const grantAdPermit = (): Promise<void> => browser.storage.local.set({ [WEB_WIDGETS_LOCAL_AD_PERMIT]: true });
 
 export const subscribeAdPermitGranted = (onGranted: EmptyFn): EmptyFn =>
-  onStorageKey<boolean>(WEB_WIDGETS_LOCAL_AD_PERMIT, granted => {
+  onStorageChanged<boolean>(WEB_WIDGETS_LOCAL_AD_PERMIT, granted => {
     if (granted) onGranted();
   });
